Create SQL_Fragmentation and Stats
The below is script that is alot more simple than Hallengren's but i find this really quick and well tailored for given servers in this given environment that reorganizes, reindex's and even updates statistics with novel conditions such as page counts to determine whether worthwhile to reindex. 

I even decided to add full scan statistic recreations as it was very quick and was just an interesting endeavour and decided to keep.